refactor: Refactor improve class files struct
클래스별로 파일 분리하여 코드 구조 개선 및 리팩토링
